move navigation elements from header to separate component
