+++ src/components/header-section/index.tsx
@@ -1,13 +1,7 @@
 import { useState } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import { useScrollLock } from '../../hooks/useScrollLock';
-import { navigationInterface } from '../../types/appInterfaces';
-
-const navigation: navigationInterface[] = [
-	{ key: 1, name: 'about', href: '#about' },
-	{ key: 2, name: 'projects', href: '#projects' },
-	{ key: 3, name: 'contact', href: '#footer' },
-];
+import Navigation from './components/Navigation';
 
 function Header() {
 	const { pathname } = useLocation();
@@ -51,27 +45,7 @@ function Header() {
 						className="hamburgerToggleBtn ph-list"
 						onClick={onToggleNavMobile}
 					/>
-					<nav
-						className={isActive ? 'nav active' : 'nav'}
-						aria-hidden={!isActive}
-					>
-						<ul className="navList">
-							{navigation.map((item) => {
-								return (
-									<li key={item.key} className="navList__items">
-										<a
-											href={item.href}
-											aria-label="Navigate to About Me Section"
-											className="hover-underline-animation"
-											onClick={onNavigate}
-										>
-											{item.name}
-										</a>
-									</li>
-								);
-							})}
-						</ul>
-					</nav>
+					<Navigation isActive={isActive} onNavigate={onNavigate} />
 				</>
 			)}
 		</header>
